pyproject.toml
[tool.poetry]
name = "pythonappiumautomation"
version = "0.1.0"
description = ""
authors = ["Ekaterina <e.gorbushina@cdek.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^7.1.2"
selenium = "^4.2.0"
Appium-Python-Client = "^2.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
